Lock pthread mutex at initialization
The phch_common::init function returns with all mutex unlocked as the
initial state. It allows multiple threads to transmit at the same time
at the begining. It causes BladeRF TX error described in issue #180 and #142.

This patch fixes this bug by locking all the mutex in the init function
except the first mutex.