fix overlapping memcpy detected by valgrind
this has been detected by valgrind+gcc on Ubuntu 16.04 but
isn't an issue for more recent versions. Not sure if that is because
of an issue in older libc and there memcpy implementation or a
bug in valgrind, or both.

In any case, overlapping memcpy may cause undefined behaviour so
it seems ok to only copy the symbols that do not contain the
DMRS, i.e. not symbol 3 in each slot.