[ENB] enb.conf.example: Mention WRITE_SIB_PCAP
The documentation talks about a MAC protocol trace, but in fact
it is only a protocol trace of dedicated channels.  There's a related
define in the source code, but that's not documented.

Let's at least make the user aware that the MAC pcap trace is
for dedicated channels only, unless he uses that #define.