zmq: add monitor code for sockets but disable it
according to the ZMQ dev guide, one can use a socket monitor
to get informed about changes to a socket, e.g. when a client disconnected.

This is useful to detect when a UE dropped the connection to reinitialize
the socket and timers in a eNB.

This commit adds code to create such a socket monitor but leaves
it disabled because it still doesn't work as expected.