benchmark_radio: added option to change sampling rate during experiment
this patch adds the -y paramter that changes the sampling rate
during a radio benchmark to stress-test the API

it turned out that some devices/drivers are very prone to
sampling rate changes. since this happens during UE operation
when the UE is searching for cells, we need to make sure
the driver/device supports that operation.