Fix linker error on GCC 10.1 due to -fno-common being default
Due to GCC 10.1 defaulting to -fno-common, global variable declarations
seem to need to be declared as extern now.

Note that the global should only be declared now in one file, and since
there's a regular and 16-bit implementation of this AVX2 code, this
patch might not fix the issue in 16 bit mode.

Compiles on GCC 10.1.0 on x86_64