enb: fix reestablishment with AS security enabled
this fixes RRC reestablishment with security turned on

the bearers, primarily SRB1, were reestablished before the "old"
security context was restored from the previous RNTI of the
reestablishing user.

this let SRB1 without proper security config. therefore the UE
couldn't decrypt the first response of the UE on SRB1 that was
already encrypted.