srsenb: Handle E-RABReleaseCommand - MME Initiated
The E-RAB RELEASE COMMAND message shall contain the information required by the eNB
to release at least one E-RAB in the E-RAB To Be Released List IE.
If a NAS-PDU IE is contained in the message, the eNB shall pass it to the UE.

Upon reception of the E-RAB RELEASE COMMAND message the eNB shall execute the
release of the requested E-RABs i.e. release of Data Radio Bearer + resources on Uu interface + S1 resource.

The eNB shall report to the MME, in the E-RAB RELEASE RESPONSE message,
the result for all the E-RABs to be released.
- A list of E-RABs which are released successfully
- A list of E-RABs which failed to be released

If the eNB receives an E-RAB RELEASE COMMAND message containing some E-RAB ID IEs that eNB does not recognize,
the eNB shall report the corresponding invalid E-RABs as failed in the E-RAB RELEASE RESPONSE message
with the appropriate cause, e.g., “Unknown E-RAB ID”