ue,phy: reduce log level when 2nd grant for same PID is registered
this resolves issue #2277 in which a warning was logged when
the CC worker decoded a 2nd phantom PDDCH.

It's difficult to avoid this 2nd PDDCH without addional computation
that we don't want add at this stage.

Rather than filtering out the Warning in the log post-processing,
we reduce the log level to info.