rlc_am_lte: fix handling of maxretx
this patch fixes the actions/handling after RLC detected
maxRetx reached for a given SN.

According to the TS, RLC should only inform upper layers and
not try to recover from the event itself.

As a consequence, we won't manipulate the Tx or Rx window.
As a result of this, we might retransmit a SN more than
the specified amount of times.

It's the task of RRC to reestablish the bearer to recover
from that.