rlc_am_lte: refactor logging when PDCP SDUs are enqueued
the patch refactor the logging when a new PDCP SDU is enqueued for
transmission at RLC.

If the SN is already present, only a warning is logged. From the RLC
perspective operation continues and the SDU will be transmitted.

The patch also changes the order of logs. When the SN cannot be inserted
inside the queue of undelivered SDUs, only one message is logged.